Trello: Check if board is owned by user
Added trello api call for getting user's own member record.
Created BoardMember class for storing membership for a board.
In CreateTrelloBoardIfNeeded, check if the board matched belongs to the user.
Added config value ownBoardsOnly which if set will disallow using other people's
boards.
If not set, then the user will be prompted as to whether or not they want to use
someone  else's board.
If they don't, a new board will be created.
Changed logging level for rest calls to DEBUG as it is useful information.