Bugzilla: Add actions for adding tracking Jira issues
Use createTrackingIssueForBug workflow for creating a tracking issue in Jira for
a  Bugzilla bug.
Use createTrackingIssuesForQuery workflow to create tracking issues in Jira for
a  named query in Bugzilla.
This allows bulk creation of tracking issues for bugs in a release query.
A comment is added to the bug for each tracking issue that is created. This is
used  to prevent adding duplicate tracking issues.