WorkflowAction: Add method for failing if condition not met
Similar to cannotRunAction, failWorkflowIfConditionNotMet will fail the workflow
if a non null string is returned.
That string is treated as the reason why the workflow is failed.
Added support to BaseAction for failing the workflow if specified commands are
missing.
Moved all parsing of changelist text into the parsing code into
ReviewRequestDraft.