Services: Treat git and perforce as services
Added creation of new instances for Git and Perforce to ServiceLocator.
Removed Perforce as a parent property. Only reference in actions that use
perforce.
Changed constructor for Perforce to use client name.
The working directory can be derived from the client name.
This change ensures that all p4 commands are executed from the client root
directory.