Perforce: Compare perforce diff to git diff
ExitIfChangelistDoesNotMatchGitBranch is used to verify that a changelist
matches a git branch.
Changed diff routine used by perforce to be git based.
Modified the content returned to exactly match a git diff.
Added diff content for added and deleted files to the perforce diff as well.
Added action SaveChangelistDiffToFile to save a git diff of a changelist to an
output file