Diffs: Handle empty file diffs and changes to files not in source control
When diffing perforce files, a newly added file could be empty.
When diffing a newly added file against a previous commit that added the file, the file won't be in
source control.
Handle this for perforce diffs by using p4 where to get the full depot path for the file.