Jenkins: Various features
Added showing of date to job test failures if the last run was a few days ago.
Changed parsing of test and config failures to be from the testNG results files directly.
This allows capturing of all config method runs not just the failures.
Added configMethod as column in test_result table.
Fixed issue with incorrectly removing dataProvider indices.